# src/aegis_range/openapi/aegis-bank.json
{
  "openapi": "3.1.0",
  "info": {"title": "Aegis Bank API", "version": "1.0.0"},
  "servers": [{"url": "http://aegis-bank:8101"}],
  "paths": {
    "/api/profile": {
      "get": {"operationId": "getCurrentProfile", "summary": "Get the current customer profile", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Profile"}, "401": {"description": "Authentication required"}}},
      "patch": {"operationId": "updateCurrentProfile", "summary": "Update the current customer profile", "security": [{"bearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "properties": {"display_name": {"type": "string", "maxLength": 80}, "timezone": {"type": "string", "maxLength": 60}}}}}}, "responses": {"200": {"description": "Updated profile"}, "401": {"description": "Authentication required"}}}
    },
...
  },
  "components": {"securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}}}
}